requirements.txt
tweepy==4.14.0
python-dotenv==1.0.0
praw==7.7.1
openai==1.12.0 